Move plugin discoverable logic into core-plugins
