+++ codex-rs/core-plugins/src/lib.rs
@@ -1,3 +1,4 @@
+mod discoverable;
 pub mod installed_marketplaces;
 pub mod loader;
 mod manager;
@@ -20,42 +21,11 @@ pub mod toggles;
 pub const OPENAI_CURATED_MARKETPLACE_NAME: &str = "openai-curated";
 pub const OPENAI_BUNDLED_MARKETPLACE_NAME: &str = "openai-bundled";
 
-pub const TOOL_SUGGEST_DISCOVERABLE_PLUGIN_ALLOWLIST: &[&str] = &[
-    "github@openai-curated",
-    "notion@openai-curated",
-    "slack@openai-curated",
-    "gmail@openai-curated",
-    "google-calendar@openai-curated",
-    "google-drive@openai-curated",
-    "openai-developers@openai-curated",
-    "canva@openai-curated",
-    "teams@openai-curated",
-    "sharepoint@openai-curated",
-    "outlook-email@openai-curated",
-    "outlook-calendar@openai-curated",
-    "linear@openai-curated",
-    "figma@openai-curated",
-    "github@openai-curated-remote",
-    "notion@openai-curated-remote",
-    "slack@openai-curated-remote",
-    "gmail@openai-curated-remote",
-    "google-calendar@openai-curated-remote",
-    "google-drive@openai-curated-remote",
-    "openai-developers@openai-curated-remote",
-    "canva@openai-curated-remote",
-    "teams@openai-curated-remote",
-    "sharepoint@openai-curated-remote",
-    "outlook-email@openai-curated-remote",
-    "outlook-calendar@openai-curated-remote",
-    "linear@openai-curated-remote",
-    "figma@openai-curated-remote",
-    "chrome@openai-bundled",
-    "computer-use@openai-bundled",
-];
-
 pub type LoadedPlugin = codex_plugin::LoadedPlugin<codex_config::McpServerConfig>;
 pub type PluginLoadOutcome = codex_plugin::PluginLoadOutcome<codex_config::McpServerConfig>;
 
+pub use discoverable::ToolSuggestDiscoverablePlugin;
+pub use discoverable::ToolSuggestPluginDiscoveryInput;
 pub use manager::ConfiguredMarketplace;
 pub use manager::ConfiguredMarketplaceListOutcome;
 pub use manager::ConfiguredMarketplacePlugin;
